Use data table for search results
