+++ src/app/components/home/home.component.html
@@ -1,1 +1,1 @@
-<app-data-table *ngIf="caseRows$ | async as rows" [rows]="rows" [columns]="columns"></app-data-table>
+<h1 class="govuk-heading-l">Welcome to DARTS</h1>
