+++ src/app/components/search/results/results.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { ResultsComponent, SortableColumn, SortingInterface } from './results.component';
+import { ResultsComponent } from './results.component';
 import { CaseData, HearingData } from '@darts-types/index';
 
 describe('ResultsComponent', () => {
@@ -163,165 +163,6 @@ describe('ResultsComponent', () => {
       expect(component.getHearingsValue).toHaveBeenCalled();
       expect(dValue).toBe('');
       expect(cValue).toBe('');
-    });
-  });
-
-  describe('Pagination', () => {
-    it('pages cases based on page limit', () => {
-      component.cases = MOCK_CASES;
-      component.currentPage = 1;
-      component.pageLimit = 1;
-
-      component.ngOnChanges();
-
-      expect(component.pagedCases.length).toEqual(1);
-    });
-
-    it('pages cases based on current page', () => {
-      // Given 4 cases and a page limit of 3 and on page 2
-      component.cases = MOCK_CASES;
-      component.pageLimit = 3;
-
-      component.onPageChanged(2);
-
-      expect(component.pagedCases.length).toEqual(1);
-      expect(component.pagedCases[0]).toEqual(MOCK_CASES[3]);
-    });
-  });
-
-  describe('Sorting', () => {
-    it('should set sorting to ascending order for the given column', () => {
-      component.cases = MOCK_CASES;
-      const column: SortableColumn = 'case_number';
-
-      component.sorting = {
-        column,
-        order: 'desc',
-      };
-
-      component.sortTable(column);
-
-      const expectedSorting: SortingInterface = {
-        column,
-        order: 'asc',
-      };
-
-      expect(component.sorting).toEqual(expectedSorting);
-      expect(component.cases).toEqual(component.cases.reverse());
-    });
-
-    it('should set sorting to descending order for the given column', () => {
-      component.cases = MOCK_CASES;
-      const column: SortableColumn = 'case_number';
-      component.sorting = {
-        column,
-        order: 'asc',
-      };
-
-      component.sortTable(column);
-
-      const expectedSorting: SortingInterface = {
-        column,
-        order: 'desc',
-      };
-
-      expect(component.sorting).toEqual(expectedSorting);
-      expect(component.cases).toEqual(component.cases);
-    });
-
-    it('should compare two strings in ascending order', () => {
-      const column: SortableColumn = 'case_number';
-      const result = component.compareStrings(column, 'abc', 'def');
-      expect(result).toBe(1); // 'abc' comes before 'def'
-    });
-
-    it('should compare two strings in descending order', () => {
-      const column: SortableColumn = 'case_number';
-      const result = component.compareStrings(column, 'def', 'abc');
-      expect(result).toBe(-1); // 'def' comes after 'abc'
-    });
-
-    it('should return true for isDescSorting when sorting in descending order', () => {
-      const column: SortableColumn = 'case_number';
-      component.sorting = {
-        column,
-        order: 'desc',
-      };
-      const isDesc = component.isDescSorting(column);
-      expect(isDesc).toBe(true);
-    });
-
-    it('should return false for isDescSorting when sorting in ascending order', () => {
-      const column: SortableColumn = 'case_number';
-      component.sorting = {
-        column,
-        order: 'asc',
-      };
-      const isDesc = component.isDescSorting(column);
-      expect(isDesc).toBe(false);
-    });
-
-    it('should return true for isAscSorting when sorting in ascending order', () => {
-      const column: SortableColumn = 'case_number';
-      component.sorting = {
-        column,
-        order: 'asc',
-      };
-      const isAsc = component.isAscSorting(column);
-      expect(isAsc).toBe(true);
-    });
-
-    it('should return false for isAscSorting when sorting in descending order', () => {
-      const column: SortableColumn = 'case_number';
-      component.sorting = {
-        column,
-        order: 'desc',
-      };
-      const isAsc = component.isAscSorting(column);
-      expect(isAsc).toBe(false);
-    });
-
-    it('should return "ascending" for getAriaSort when sorting in ascending order', () => {
-      const column: SortableColumn = 'case_number';
-      component.sorting = {
-        column,
-        order: 'asc',
-      };
-      const ariaSort = component.getAriaSort(column);
-      expect(ariaSort).toBe('ascending');
-    });
-
-    it('should return "descending" for getAriaSort when sorting in descending order', () => {
-      const column: SortableColumn = 'case_number';
-      component.sorting = {
-        column,
-        order: 'desc',
-      };
-      const ariaSort = component.getAriaSort(column);
-      expect(ariaSort).toBe('descending');
-    });
-
-    it('should return "none" for getAriaSort when not sorting the column', () => {
-      const column: SortableColumn = 'case_number';
-      component.sorting = {
-        column: 'other_column', // A different column
-        order: 'asc',
-      };
-      const ariaSort = component.getAriaSort(column);
-      expect(ariaSort).toBe('none');
-    });
-
-    it('should reset column sorting state on cases changed', () => {
-      component.cases = MOCK_CASES;
-      const column: SortableColumn = 'case_number';
-      component.sorting = {
-        column,
-        order: 'asc',
-      };
-
-      component.ngOnChanges();
-
-      expect(component.sorting.column).toBe('');
     });
   });
 });
